requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.4
Flask-Mail==0.9.1
bcrypt==4.0.1
PyJWT==2.8.0
requests==2.31.0
